+++ GatorSorter/arrayDivider.h
@@ -7,7 +7,7 @@ using namespace std;
 
 class arrayDivider {
 public:
-    arrayDivider(char const *path, int sampling_frequency, int64_t ram_limit);
+    arrayDivider(char const *path, int64_t sampling_frequency, int64_t ram_limit);
     arrayDivider(const arrayDivider& orig);
 	int data_type = 0;
 	int num_bytes_per_entry = 0;
